refactor redux folder structure
