+++ src/App.js
@@ -10,7 +10,7 @@ import {
 } from "./App.styles";
 import { Coin, Home, Portfolio } from "pages";
 import { Currency, GlobalInfo, SearchCoin } from "components";
-import { setCurrency } from "store/app/appActions";
+import { setCurrency } from "store/app/actions";
 
 console.clear();
