+++ src/store/home/actions.js
@@ -1,18 +1,16 @@
 import axios from "axios";
 import queryString from "query-string";
 import { keysToSnakeCase, keysToCamelCase } from "utils";
-
-export const PARSE_HOME_QUERY_STRING = "PARSE_HOME_QUERY_STRING";
-
-export const FETCH_ALL_COINS_SUCCESS = "FETCH_ALL_COINS_SUCCESS";
-export const FETCH_ALL_COINS_PENDING = "FETCH_ALL_COINS_PENDING";
-export const FETCH_ALL_COINS_ERROR = "FETCH_ALL_COINS_ERROR";
-
-export const FETCH_PRICES_PENDING = "FETCH_PRICES_PENDING";
-export const FETCH_PRICES_SUCCESS = "FETCH_PRICES_SUCCESS";
-export const FETCH_PRICES_ERROR = "FETCH_PRICES_ERROR";
-
-export const TOGGLE_ORDER = "TOGGLE_ORDER";
+import {
+  PARSE_HOME_QUERY_STRING,
+  FETCH_ALL_COINS_SUCCESS,
+  FETCH_ALL_COINS_PENDING,
+  FETCH_ALL_COINS_ERROR,
+  FETCH_PRICES_PENDING,
+  FETCH_PRICES_SUCCESS,
+  FETCH_PRICES_ERROR,
+  TOGGLE_ORDER,
+} from "./index";
 
 export const fetchAllCoins = () => async (dispatch, getState) => {
   try {
